requirements.txt
gm>=3.0.180
numpy<2.0
pandas<2.0
python-dateutil>=2.8.1
python-dotenv>=1.0.0
